Upgrade the task
